Creation of dev branch - edit readme to match branch + small changes to Endpoints

+++ P90's Twitch Library/API/Endpoints/CreateCustomRewards.cs
@@ -35,7 +35,7 @@ namespace P90Ez.Twitch.API.Endpoints
         /// <strong>Creates a Custom Channelpoint Reward in the broadcaster’s channel.</strong>
         /// <para>Scope: <em>channel:manage:redemptions</em></para>
         /// <para>TokenType: <em>User Access Token</em></para>
-        /// <see href="https://dev.twitch.tv/docs/api/reference#start-commercial">Twitch Docs (Error Handling)</see> 
+        /// <see href="https://dev.twitch.tv/docs/api/reference/#create-custom-rewards">Twitch Docs (Error Handling)</see> 
         /// </summary>
         /// <param name="credentials">Your generated login credentials</param>
         /// <param name="broadcaster_id">The ID of the broadcaster to add the custom reward to.</param>
@@ -105,7 +105,7 @@ namespace P90Ez.Twitch.API.Endpoints
             /// <summary>
             /// <em>[OPTIONAL]</em> The maximum number of redemptions allowed per live stream. Applied only if is_max_per_stream_enabled is true. The minimum value is 1.
             /// </summary>
-            public int? max_per_stream { get; set; }
+            public long? max_per_stream { get; set; }
             /// <summary>
             /// <em>[OPTIONAL]</em> A Boolean value that determines whether to limit the maximum number of redemptions allowed per user per stream (see the max_per_user_per_stream field). The default is false.
             /// </summary>
@@ -113,7 +113,7 @@ namespace P90Ez.Twitch.API.Endpoints
             /// <summary>
             /// <em>[OPTIONAL]</em> The maximum number of redemptions allowed per user per stream. Applied only if is_max_per_user_per_stream_enabled is true. The minimum value is 1.
             /// </summary>
-            public int? max_per_user_per_stream { get; set; }
+            public long? max_per_user_per_stream { get; set; }
             /// <summary>
             /// <em>[OPTIONAL]</em> A Boolean value that determines whether to apply a cooldown period between redemptions (see the global_cooldown_seconds field for the duration of the cooldown period). The default is false.
             /// </summary>
@@ -121,7 +121,7 @@ namespace P90Ez.Twitch.API.Endpoints
             /// <summary>
             /// <em>[OPTIONAL]</em> The cooldown period, in seconds. Applied only if the is_global_cooldown_enabled field is true. The minimum value is 1; however, the minimum value is 60 for it to be shown in the Twitch UX.
             /// </summary>
-            public int? global_cooldown_seconds { get; set; }
+            public long? global_cooldown_seconds { get; set; }
             /// <summary>
             /// <em>[OPTIONAL]</em> A Boolean value that determines whether redemptions should be set to FULFILLED status immediately when a reward is redeemed. If false, status is set to UNFULFILLED and follows the normal request queue process. The default is false.
             /// </summary>
